+++ gpu_bdb/benchmark_runner/benchmark_config.yaml
@@ -6,7 +6,7 @@ file_format: parquet
 output_filetype: parquet
 split_row_groups: False
 repartition_small_table: True
-benchmark_runner_include_bsql:
+benchmark_runner_include_sql:
 
 scheduler_file_path:
 dask_profile: False
