+++ gpu_bdb/queries/q04/gpu_bdb_query_04.py
@@ -16,15 +16,20 @@
 
 import sys
 
-
 from bdb_tools.utils import (
     benchmark,
     gpubdb_argparser,
     run_query,
 )
+
 from bdb_tools.readers import build_reader
+
 from bdb_tools.sessionization import get_sessions
 
+from bdb_tools.q04_utils import (
+    abandonedShoppingCarts,
+    reduction_function
+)
 
 # parameters
 q04_session_timeout_inSec = 3600
@@ -50,60 +55,6 @@ def read_tables(config):
     web_clicksteams_df = table_reader.read("web_clickstreams", relevant_cols=wcs_cols)
 
     return wp_df, web_clicksteams_df
-
-
-def abandonedShoppingCarts(df, DYNAMIC_CAT_CODE, ORDER_CAT_CODE):
-    import cudf
-
-    # TODO: test without reset index
-    df.reset_index(drop=True, inplace=True)
-
-    # Select groups where last dynamic row comes after last order row
-    filtered_df = df[
-        (df["wp_type_codes"] == ORDER_CAT_CODE)
-        | (df["wp_type_codes"] == DYNAMIC_CAT_CODE)
-    ]
-    # TODO: test without reset index
-    filtered_df.reset_index(drop=True, inplace=True)
-    # Create a new column that is the concatenation of timestamp and wp_type_codes
-    # (eg:123456:3, 234567:5)
-    filtered_df["wp_type_codes"] = (
-        filtered_df["tstamp_inSec"]
-        .astype("str")
-        .str.cat(filtered_df["wp_type_codes"].astype("str"), sep=":")
-    )
-    # This gives the last occurrence (by timestamp) within the "order", "dynamic" wp_types
-    filtered_df = filtered_df.groupby(
-        ["wcs_user_sk", "session_id"], as_index=False, sort=False
-    ).agg({"wp_type_codes": "max"})
-    # If the max contains dynamic, keep the row else discard.
-    last_dynamic_df = filtered_df[
-        filtered_df["wp_type_codes"].str.contains(
-            ":" + str(DYNAMIC_CAT_CODE), regex=False
-        )
-    ]
-    del filtered_df
-
-    # Find counts for each group
-    grouped_count_df = df.groupby(
-        ["wcs_user_sk", "session_id"], as_index=False, sort=False
-    ).agg({"tstamp_inSec": "count"})
-    # Merge counts with the "dynamic" shopping cart groups
-    result = last_dynamic_df.merge(
-        grouped_count_df, on=["wcs_user_sk", "session_id"], how="inner"
-    )
-    del (last_dynamic_df, grouped_count_df)
-    return cudf.DataFrame(
-        {"pagecount": result.tstamp_inSec.sum(), "count": len(result)}
-    )
-
-
-def reduction_function(df, keep_cols, DYNAMIC_CAT_CODE, ORDER_CAT_CODE):
-    df = get_sessions(df, keep_cols=keep_cols)
-    df = abandonedShoppingCarts(
-        df, DYNAMIC_CAT_CODE=DYNAMIC_CAT_CODE, ORDER_CAT_CODE=ORDER_CAT_CODE
-    )
-    return df
 
 
 def main(client, config):
